+++ src/f1_robot_model/launch/amcl.launch.py
@@ -5,30 +5,49 @@ from launch_ros.actions import Node
 import os
 from launch.actions import DeclareLaunchArgument
 from ament_index_python.packages import get_package_share_directory
+from nav2_common.launch import RewrittenYaml
 
 def generate_launch_description():
-    # pkg name and pkg path
-    package_name = "f1_robot_model"
+    # pkg path
     pkg_share = launch_ros.substitutions.FindPackageShare(package='f1_robot_model').find('f1_robot_model')
-    # model path and config, world path
+    # robot model path
     default_model_path = os.path.join(pkg_share, 'urdf/racecar.urdf')
-    default_rviz_config_path = os.path.join(pkg_share, 'rviz/urdf_config.rviz')
-    world_path=os.path.join(pkg_share, 'world/demomap/model.sdf')
-    # use sim time? or not
+    # rviz config for sim
+    default_rviz_config_path = os.path.join(pkg_share, 'rviz/sim_config.rviz')
+    # gazebo world model path
+    world_path=os.path.join(pkg_share, 'world/demomap_2/model.sdf')
+    # use_sim_time param
     use_sim_time = LaunchConfiguration('use_sim_time')
-
-    # cartographer setting file 1
-    cartographer_config_dir = LaunchConfiguration('cartographer_config_dir',
-                                                default=os.path.join(get_package_share_directory(package_name) , 'config'))
-    # cartographer setting file 2
-    configuration_basename = LaunchConfiguration('configuration_basename', default='localization.lua')
-    pbstream_file = LaunchConfiguration('pbstream_dir',
-                                                default=os.path.join(get_package_share_directory(package_name) , 'map', 'testmap.pbstream'))
-    resolution = LaunchConfiguration('resolution', default='0.05')
-    publish_period_sec = LaunchConfiguration('publish_period_sec', default='0.5')
+    # namspace declare
+    namespace = LaunchConfiguration('namespace')
+    # node respawn param for life cycle node
+    use_respawn = LaunchConfiguration('use_respawn')
+    # node list for life cycle node
+    lifecycle_nodes = ['map_server', 'amcl']
+    # autostart setting for nav2 stack (life cycle node)
+    autostart = LaunchConfiguration('autostart')
+    # param files and map file for amcl and map server
+    params_file = LaunchConfiguration('params_file')
+    map_yaml_file = LaunchConfiguration('map')
+    map_dir = os.path.join(pkg_share, 'map')
+    map_file = 'demomap_2.yaml'
+    param_dir = os.path.join(pkg_share, 'config')
+    param_file = 'amcl.yaml'
 
     # joy_teleop config file path
     joy_teleop_config_file = os.path.join(pkg_share, 'config/MXswitch.config.yaml')
+
+    param_substitutions = {
+        'use_sim_time': use_sim_time,
+        'yaml_filename': map_yaml_file
+    }
+
+    configured_params = RewrittenYaml(
+        source_file=params_file,
+        root_key=namespace,
+        param_rewrites=param_substitutions,
+        convert_types=True
+    )
 
     # Position and orientation
     # [X, Y, Z]
@@ -36,11 +55,35 @@ def generate_launch_description():
     # [Roll, Pitch, Yaw]
     orientation = [0.0, 0.0, 0.0]
 
+    declare_namespace_cmd = DeclareLaunchArgument(
+        'namespace',
+        default_value='',
+        description='Top-level namespace')
+
+    declare_use_respawn_cmd = DeclareLaunchArgument(
+        'use_respawn', default_value='False',
+        description='Whether to respawn if a node crashes. Applied when composition is disabled.'
+        )
+
     declare_use_sim_time_cmd = DeclareLaunchArgument(
         name='use_sim_time',
         default_value='true',
         description='Use simulation (Gazebo) clock if true'
         )
+
+    declare_autostart_cmd = DeclareLaunchArgument(
+        'autostart', default_value='true',
+        description='Automatically startup the nav2 stack')
+
+    declare_map_yaml_cmd = DeclareLaunchArgument(
+        'map',
+        default_value=os.path.join(map_dir, map_file),
+        description='[localize] Full path to map yaml file to load')
+
+    declare_params_file_cmd = DeclareLaunchArgument(
+        'params_file',
+        default_value=os.path.join(param_dir, param_file),
+        description='Full path to the ROS2 parameters file to use')
 
     robot_state_publisher_node = Node(
         package='robot_state_publisher',
@@ -77,34 +120,34 @@ def generate_launch_description():
         output='screen'
     )
 
-    cartographer = Node(
-        package='cartographer_ros',
-        executable='cartographer_node',
-        name='cartographer_node',
+    map_server = Node(
+        package='nav2_map_server',
+        executable='map_server',
+        name='map_server',
         output='screen',
-        parameters=[{'use_sim_time': use_sim_time}],
-        arguments=[
-            '-configuration_directory', cartographer_config_dir,
-            '-configuration_basename', configuration_basename,
-            '-load_state_filename', pbstream_file
-            ]
-    )
-    # Executing Cartographer
-    cartographer_grid = Node(
-        package='cartographer_ros',
-        executable='cartographer_occupancy_grid_node',
-        name='occupancy_grid_node',
-        output='screen',
-        parameters=[{'use_sim_time': use_sim_time}],
-        arguments=['-resolution', resolution, '-publish_period_sec', publish_period_sec]
+        respawn=use_respawn,
+        respawn_delay=2.0,
+        parameters=[configured_params],
     )
 
-    trajectory_to_odom = Node(
-        package='cartographer_ros',
-        executable='trajectory_to_odom',
-        name='trajectory_to_odom_node',
+    amcl = Node(
+        package='nav2_amcl',
+        executable='amcl',
+        name='amcl',
         output='screen',
-        parameters=[{'use_sim_time': use_sim_time}],
+        respawn=use_respawn,
+        respawn_delay=2.0,
+        parameters=[configured_params],
+    )
+
+    lifecycle_manager = Node(
+        package='nav2_lifecycle_manager',
+        executable='lifecycle_manager',
+        name='lifecycle_manager_localization',
+        output='screen',
+        parameters=[{'use_sim_time': use_sim_time},
+                    {'autostart': autostart},
+                    {'node_names': lifecycle_nodes}],
     )
 
     joy = Node(
@@ -132,15 +175,20 @@ def generate_launch_description():
         launch.actions.DeclareLaunchArgument(name='rvizconfig', default_value=default_rviz_config_path,
                                             description='Absolute path to rviz config file'),
         launch.actions.ExecuteProcess(cmd=['gazebo', '--verbose', '-s', 'libgazebo_ros_init.so', '-s', 'libgazebo_ros_factory.so', world_path], output='screen'),
+        declare_namespace_cmd,
         declare_use_sim_time_cmd,
+        declare_use_respawn_cmd,
+        declare_autostart_cmd,
+        declare_params_file_cmd,
+        declare_map_yaml_cmd,
         joint_state_publisher_node,
         robot_state_publisher_node,
         spawn_entity,
         rviz_node,
         ackermann_to_twist_converter_node,
-        cartographer,
-        cartographer_grid,
-        trajectory_to_odom,
+        amcl,
+        map_server,
+        lifecycle_manager,
         joy,
         joy_teleop,
     ])