+++ src/f1_robot_model/launch/display.launch.py
@@ -7,10 +7,13 @@ from launch.actions import DeclareLaunchArgument
 from ament_index_python.packages import get_package_share_directory
 
 def generate_launch_description():
+    # pkg path
     pkg_share = launch_ros.substitutions.FindPackageShare(package='f1_robot_model').find('f1_robot_model')
+    # model path and world, config path
     default_model_path = os.path.join(pkg_share, 'urdf/racecar.urdf')
     default_rviz_config_path = os.path.join(pkg_share, 'rviz/urdf_config.rviz')
-    world_path=os.path.join(pkg_share, 'world/demomap/model.sdf')
+    world_path=os.path.join(pkg_share, 'world/demomap_2/model.sdf')
+    # use sim time or not?
     use_sim_time = LaunchConfiguration('use_sim_time')
 
     # Position and orientation
@@ -18,6 +21,9 @@ def generate_launch_description():
     position = [0.0, 0.0, 0.0]
     # [Roll, Pitch, Yaw]
     orientation = [0.0, 0.0, 0.0]
+
+    # joy_teleop config file path
+    joy_teleop_config_file = os.path.join(pkg_share, 'config/MXswitch.config.yaml')
 
     declare_use_sim_time_cmd = DeclareLaunchArgument(
         name='use_sim_time',
@@ -60,6 +66,24 @@ def generate_launch_description():
         output='screen'
     )
 
+    joy = Node(
+        package='joy',
+        executable='joy_node',
+        name='joy_node',
+        parameters=[{
+            'device_id': 0,          # ls /dev/input/js* to find the device id
+            'deadzone': 0.3,         # Deadzone for the joystick axes
+            'autorepeat_rate': 20.0, # Autorepeat rate for the joystick buttons
+        }]
+    )
+
+    joy_teleop = Node(
+        package='teleop_twist_joy',
+        executable='teleop_node',
+        name='teleop_twist_joy_node',
+        parameters=[joy_teleop_config_file]
+    )
+
     return launch.LaunchDescription([
 
         launch.actions.DeclareLaunchArgument(name='model', default_value=default_model_path,
@@ -73,4 +97,6 @@ def generate_launch_description():
         spawn_entity,
         rviz_node,
         ackermann_to_twist_converter_node,
+        joy,
+        joy_teleop,
     ])